requirements.txt
# LiveKit dependencies
livekit-agents[google]~=1.0
livekit-cli>=1.0.0
python-dotenv>=1.0.0

# Supabase dependencies
supabase>=2.7.0
postgrest>=0.16.0

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Additional utilities
asyncpg>=0.29.0
python-dateutil>=2.8.0
pytz>=2024.1

# Development dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0 